+++ src/Flutter/Widgets/Gestures.cs
@@ -13,6 +13,8 @@ public sealed class Listener : SingleChildRenderObjectWidget
         Widget? child = null,
         Action<PointerDownEvent>? onPointerDown = null,
         Action<PointerMoveEvent>? onPointerMove = null,
+        Action<PointerEnterEvent>? onPointerEnter = null,
+        Action<PointerExitEvent>? onPointerExit = null,
         Action<PointerHoverEvent>? onPointerHover = null,
         Action<PointerUpEvent>? onPointerUp = null,
         Action<PointerCancelEvent>? onPointerCancel = null,
@@ -22,6 +24,8 @@ public sealed class Listener : SingleChildRenderObjectWidget
     {
         OnPointerDown = onPointerDown;
         OnPointerMove = onPointerMove;
+        OnPointerEnter = onPointerEnter;
+        OnPointerExit = onPointerExit;
         OnPointerHover = onPointerHover;
         OnPointerUp = onPointerUp;
         OnPointerCancel = onPointerCancel;
@@ -32,6 +36,10 @@ public sealed class Listener : SingleChildRenderObjectWidget
     public Action<PointerDownEvent>? OnPointerDown { get; }
 
     public Action<PointerMoveEvent>? OnPointerMove { get; }
+
+    public Action<PointerEnterEvent>? OnPointerEnter { get; }
+
+    public Action<PointerExitEvent>? OnPointerExit { get; }
 
     public Action<PointerHoverEvent>? OnPointerHover { get; }
 
@@ -48,6 +56,8 @@ public sealed class Listener : SingleChildRenderObjectWidget
         return new RenderPointerListener(
             onPointerDown: OnPointerDown,
             onPointerMove: OnPointerMove,
+            onPointerEnter: OnPointerEnter,
+            onPointerExit: OnPointerExit,
             onPointerHover: OnPointerHover,
             onPointerUp: OnPointerUp,
             onPointerCancel: OnPointerCancel,
@@ -60,6 +70,8 @@ public sealed class Listener : SingleChildRenderObjectWidget
         var listener = (RenderPointerListener)renderObject;
         listener.OnPointerDown = OnPointerDown;
         listener.OnPointerMove = OnPointerMove;
+        listener.OnPointerEnter = OnPointerEnter;
+        listener.OnPointerExit = OnPointerExit;
         listener.OnPointerHover = OnPointerHover;
         listener.OnPointerUp = OnPointerUp;
         listener.OnPointerCancel = OnPointerCancel;